ai-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
ollama>=0.3.0
python-multipart>=0.0.6
scikit-learn>=1.3.0
numpy>=1.24.0
openai-whisper>=20231117
# Optional: speaker diarization (set HUGGINGFACE_TOKEN). Without token, transcribe-with-speakers returns text only.
pyannote-audio>=3.0.0
torch>=2.0.0
